Project-Wellify/requirements.txt
numpy
pandas
flask
pillow>=9.0.0,<10.0.0
opencv-python
pytesseract
scikit-learn
tensorflow>=2.13.0
tf-keras
torch
torchvision
transformers
spacy
nltk
pyPDF2
pdfminer.six
python-dotenv
gunicorn
streamlit
pyinstaller